requirements.txt
torch>=1.10.0
torchvision>=0.11.0
opencv-python>=4.5.0
pyserial>=3.5
unitree_sdk2py>=2.0.0
numpy>=1.21.0
